daily-quiz-bot/requirements.txt
APScheduler==3.9.1.post1
astroid==2.12.13
boto3==1.26.44
botocore==1.29.44
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.6
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
flask-swagger-ui==4.11.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.6.0
pylint==2.15.8
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
